refactor: change author component props
