+++ src/components/in-page/Post/aside/author.module.less
@@ -1,12 +1,9 @@
-.infor {
-  width: 60%;
+.introduce {
+  color: var(--juejin-font-1);
+  font-size: 1.167rem;
+  font-weight: 400;
+  line-height: 25px;
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
-  padding-left: 1rem;
-}
-.container {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
 }
