+++ src/components/widgets/Author/index.module.less
@@ -18,14 +18,6 @@
       font-weight: 500;
     }
 
-    .introduce {
-      color: var(--juejin-font-1);
-      font-size: 1.167rem;
-      font-weight: 400;
-      line-height: 25px;
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
+
   }
 }
